# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Photobooth Studio Sanctum API",
    "description": "Dokumentasi OpenAPI Interaktif & Testing Sandbox untuk Aplikasi Mobile Photobooth Studio (Flutter / iOS / Android) terintegrasi Laravel Sanctum.",
    "version": "1.0.0",
    "contact": {
      "name": "Photobooth Studio Dev Team",
      "email": "dev@photobooth.studio"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum Token"
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Autentikasi Aplikasi Mobile & Terbitkan Sanctum Bearer Token",
        "description": "Mengirimkan email dan password untuk mendapatkan Sanctum Bearer Token yang digunakan dalam autentikasi API aplikasi mobile.",
        "operationId": "mobileLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "mobile.dev@photobooth.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password"
                  },
                  "device_name": {
                    "type": "string",
                    "example": "iPhone 15 Pro Flutter Client"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autentikasi Berhasil",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Sanctum bearer token generated successfully",
                  "token": "1|sanctum_bearer_token_abc8891xyz",
                  "user": {
                    "name": "Mobile Developer User",
                    "email": "mobile.dev@photobooth.com",
                    "role": "mobile_client"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/booths": {
      "get": {
        "summary": "Daftar Perangkat Photobooth Aktif",
        "description": "Mengambil daftar seluruh perangkat booth yang tersedia di lokasi event.",
        "operationId": "getBooths",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar Perangkat Booth Berhasil Diambil",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "data": [
                    {
                      "id": "BOOTH-01",
                      "name": "Main Hall Booth",
                      "status": "online",
                      "location": "Zone A"
                    },
                    {
                      "id": "BOOTH-02",
                      "name": "VIP Stage Booth",
                      "status": "online",
                      "location": "Zone B"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/sessions/create": {
      "post": {
        "summary": "Memicu Sesi Foto Photobooth via Mobile QR",
        "description": "Memulai sesi pengerjaan photobooth langsung dari aplikasi mobile.",
        "operationId": "createSession",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "booth_id": {
                    "type": "string",
                    "example": "BOOTH-01"
                  },
                  "frame_id": {
                    "type": "string",
                    "example": "cyberpunk-neon-4strip"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesi Berhasil Dibuat",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Photobooth session triggered via Sanctum API",
                  "session_id": "SESH-8891",
                  "qr_code_url": "https://photobooth.studio/share/qr/session8891"
                }
              }
            }
          }
        }
      }
    },
    "/device/pair": {
      "post": {
        "summary": "Pasangkan Perangkat Booth & Terbitkan Device Token",
        "description": "Perangkat booth menggunakan kode pairing (PB-XXXX-XXXX) untuk pairing dan menerima Sanctum token yang diikat ke perangkat. Token memiliki ability device:heartbeat, device:config, device:voucher, dan device:session. Token yang terbit sebelum Phase 6 tidak memiliki device:session dan harus dicabut lalu di-pair ulang. Status: 422 kode/platfom salah, 403 perangkat dicabut, 409 sudah dipairing.",
        "operationId": "pairDevice",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "device_code": {
                    "type": "string",
                    "example": "PB-K2M7-9QXW"
                  },
                  "platform": {
                    "type": "string",
                    "enum": ["android", "windows"],
                    "example": "android"
                  },
                  "app_version": {
                    "type": "string",
                    "example": "1.2.0"
                  },
                  "device_identifier": {
                    "type": "string",
                    "example": "SN-ABC123"
                  }
                },
                "required": ["device_code", "platform"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pairing Berhasil",
            "content": {
              "application/json": {
                "example": {
                  "message": "Pairing berhasil.",
                  "token": "3|device_token_xyz",
                  "token_type": "Bearer",
                  "device": {
                    "id": 1,
                    "name": "Mall Android Booth",
                    "platform": "android",
                    "pair_status": "paired",
                    "status": "online",
                    "last_seen_at": "2026-09-12T04:00:00+00:00",
                    "app_version": "1.2.0",
                    "project": null
                  }
                }
              }
            }
          }
        }
      }
    },
    "/device/heartbeat": {
      "post": {
        "summary": "Heartbeat Perangkat (Presence & Versi)",
        "description": "Dikirim berkala oleh perangkat untuk memperbarui last_seen_at/status online dan versi aplikasi. Tidak menulis ke riwayat event kecuali versi aplikasi berubah.",
        "operationId": "deviceHeartbeat",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "app_version": {
                    "type": "string",
                    "example": "1.3.1"
                  },
                  "device_identifier": {
                    "type": "string",
                    "example": "SN-ABC123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Heartbeat Diterima",
            "content": {
              "application/json": {
                "example": {
                  "device": {
                    "id": 1,
                    "name": "Mall Android Booth",
                    "platform": "android",
                    "pair_status": "paired",
                    "status": "online",
                    "app_version": "1.3.1",
                    "project": {
                      "id": 2,
                      "name": "Mall Photobox"
                    }
                  },
                  "server_time": "2026-09-12T04:05:00+00:00"
                }
              }
            }
          }
        }
      }
    },
    "/device/config": {
      "get": {
        "summary": "Konfigurasi Runtime Perangkat",
        "description": "Mengambil proyek aktif beserta pengalaman foto (timer, layout, frame, filter, brightness) untuk perangkat. runnable=false saat proyek draf/nonaktif atau belum ditugaskan.",
        "operationId": "deviceConfig",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Konfigurasi Berhasil Diambil",
            "content": {
              "application/json": {
                "example": {
                  "device": {
                    "id": 1,
                    "name": "Mall Android Booth",
                    "platform": "android"
                  },
                  "project": {
                    "id": 2,
                    "name": "Mall Photobox",
                    "type": "retail",
                    "orientation": "portrait",
                    "welcome_image_url": "http://localhost:8000/storage/welcome/mall.jpg",
                    "status": "active",
                    "available": true
                  },
                  "experience": {
                    "timer_seconds": 12,
                    "layout": "grid_4",
                    "frame": "classic",
                    "filter": "bw",
                    "brightness": 15
                  },
                  "runnable": true,
                  "server_time": "2026-09-12T04:05:00+00:00"
                }
              }
            }
          }
        }
      }
    },
    "/voucher/validate": {
      "post": {
        "summary": "Validasi Voucher Perangkat Booth",
        "description": "Memvalidasi kode voucher terhadap perangkat yang sedang assigned ke sebuah proyek. Endpoint ini HANYA mengecek aturan bisnis dan TIDAK melakukan redeem: used_count tidak bertambah, last_used_at tidak diubah, dan tidak ada booth session yang dibuat. Untuk memakai voucher, panggil POST /session/start. Business invalidity dikembalikan dengan HTTP 200 dan reason machine-readable; kegagalan autentikasi tetap 401/403.",
        "operationId": "validateVoucher",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "VCR-7K4F-92QM",
                    "description": "Kode voucher, dinormalisasi otomatis (trim + uppercase) di sisi server."
                  }
                },
                "required": ["code"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hasil validasi (valid atau alasan bisnis tidak valid)",
            "content": {
              "application/json": {
                "examples": {
                  "valid": {
                    "summary": "Voucher valid untuk proyek perangkat",
                    "value": {
                      "valid": true,
                      "voucher": {
                        "code": "VCR-7K4F-92QM",
                        "remaining_uses": 1,
                        "expires_at": "2026-12-31",
                        "project_id": 2
                      },
                      "project": {
                        "id": 2,
                        "name": "Mall Photobox"
                      }
                    }
                  },
                  "invalid": {
                    "summary": "Voucher kadaluarsa",
                    "value": {
                      "valid": false,
                      "reason": "expired"
                    }
                  },
                  "notFound": {
                    "summary": "Kode tidak ditemukan",
                    "value": {
                      "valid": false,
                      "reason": "not_found"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada atau bukan token perangkat"
          },
          "403": {
            "description": "Token tidak memiliki ability device:voucher atau perangkat telah dicabut"
          },
          "422": {
            "description": "Field code wajib diisi"
          }
        }
      }
    },
    "/session/start": {
      "post": {
        "summary": "Mulai Sesi Booth & Redeem Voucher",
        "description": "Satu-satunya endpoint yang koneksi pemakaian voucher. Membutuhkan ability device:session. Operasi berjalan dalam SATU transaksi dengan urutan lock: baris device dikunci lebih dulu untuk menserialisasi permintaan dari booth yang sama, lalu pemeriksaan sesi aktif dilakukan, baru baris voucher dikunci dan diredeem. Jika permintaan kedua datang saat sesi masih aktif, alasan active_session_exists dikembalikan BESERTA sesi yang sedang berjalan dan pemakaian voucher TIDAK bertambah. Pembatalan sesi TIDAK mengembalikan pemakaian voucher. Kegagalan bisnis dikembalikan dengan HTTP 200/201 dan reason machine-readable; kegagalan autentikasi tetap 401/403.",
        "operationId": "startSession",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "voucher_code": {
                    "type": "string",
                    "example": "VCR-7K4F-92QM",
                    "description": "Kode voucher, dinormalisasi otomatis (trim + uppercase) di sisi server. Wajib untuk mode self_service."
                  },
                  "mode": {
                    "type": "string",
                    "enum": ["self_service", "operator"],
                    "default": "self_service",
                    "description": "self_service (default). Mode operator belum didukung pada MVP dan akan ditolak dengan unsupported_mode."
                  }
                },
                "required": ["voucher_code"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sesi berhasil dimulai dan satu pemakaian voucher telah dikurangi",
            "content": {
              "application/json": {
                "example": {
                  "valid": true,
                  "session": {
                    "id": 1,
                    "code": "SES-K2M7-9QXW",
                    "status": "active",
                    "mode": "self_service",
                    "started_at": "2026-09-13T04:10:00+00:00"
                  },
                  "project": {
                    "id": 2,
                    "name": "Mall Photobox",
                    "type": "retail",
                    "orientation": "portrait",
                    "welcome_image_url": "http://localhost:8000/storage/welcome/mall.jpg",
                    "status": "active",
                    "available": true
                  },
                  "experience": {
                    "timer_seconds": 5,
                    "layout": "strip_4",
                    "frame": "film_strip",
                    "filter": "original",
                    "brightness": 0
                  },
                  "voucher": {
                    "remaining_uses": 0
                  }
                }
              }
            }
          },
          "200": {
            "description": "Kegagalan bisnis (reason machine-readable)",
            "content": {
              "application/json": {
                "examples": {
                  "activeSessionExists": {
                    "summary": "Perangkat masih memiliki sesi aktif",
                    "value": {
                      "valid": false,
                      "reason": "active_session_exists",
                      "session": {
                        "id": 1,
                        "code": "SES-K2M7-9QXW",
                        "status": "active",
                        "mode": "self_service",
                        "started_at": "2026-09-13T04:10:00+00:00"
                      }
                    }
                  },
                  "unsupportedMode": {
                    "summary": "Mode operator belum didukung",
                    "value": {
                      "valid": false,
                      "reason": "unsupported_mode",
                      "session": null
                    }
                  },
                  "voucherInvalid": {
                    "summary": "Voucher tidak valid untuk proyek perangkat",
                    "value": {
                      "valid": false,
                      "reason": "project_mismatch",
                      "session": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada atau bukan token perangkat"
          },
          "403": {
            "description": "Token tidak memiliki ability device:session atau perangkat telah dicabut"
          },
          "404": {
            "description": "Sesi tidak ditemukan"
          },
          "422": {
            "description": "Field voucher_code wajib diisi untuk mode self_service"
          }
        }
      }
    },
    "/session/current": {
      "get": {
        "summary": "Sesi Booth yang Sedang Berjalan (Reload Recovery)",
        "description": "Read-only dan idempotent, aman dipanggil booth saat reload atau setelah crash. Hanya mengembalikan sesi milik perangkat pemanggil dengan status active. Jika tidak ada, session bernilai null. Tidak ada efek samping dan tidak ada voucher yang dikunci.",
        "operationId": "currentSession",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sesi berhasil diambil, atau null bila tidak ada",
            "content": {
              "application/json": {
                "examples": {
                  "active": {
                    "summary": "Masih ada sesi aktif",
                    "value": {
                      "valid": true,
                      "session": {
                        "id": 1,
                        "code": "SES-K2M7-9QXW",
                        "status": "active",
                        "mode": "self_service",
                        "started_at": "2026-09-13T04:10:00+00:00"
                      },
                      "project": {
                        "id": 2,
                        "name": "Mall Photobox",
                        "type": "retail",
                        "orientation": "portrait",
                        "welcome_image_url": "http://localhost:8000/storage/welcome/mall.jpg",
                        "status": "active",
                        "available": true
                      },
                      "experience": {
                        "timer_seconds": 5,
                        "layout": "strip_4",
                        "frame": "film_strip",
                        "filter": "original",
                        "brightness": 0
                      },
                      "voucher": null
                    }
                  },
                  "none": {
                    "summary": "Tidak ada sesi aktif",
                    "value": {
                      "valid": true,
                      "session": null,
                      "project": null,
                      "experience": null,
                      "voucher": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada atau bukan token perangkat"
          },
          "403": {
            "description": "Token tidak memiliki ability device:session atau perangkat telah dicabut"
          }
        }
      }
    },
    "/session/{session}/cancel": {
      "post": {
        "summary": "Batalkan Sesi Booth",
        "description": "Hanya perangkat pemilik sesi yang dapat membatalkan (diperiksa ulang di server terhadap device_id, bukan dipercaya dari id di route; perangkat lain mendapat 404). Memakai voucher TIDAK dikembalikan: voucher sudah berhasil memulai sesi dan kebijakan refund adalah keputusan fase berikutnya. Membatalkan sesi yang sudah dibatalkan aman dan idempotent, sedangkan sesi yang sudah completed tidak dikonversi kembali menjadi cancelled.",
        "operationId": "cancelSession",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "description": "ID numerik booth session",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sesi dibatalkan, atau sudah dalam status final (idempotent)",
            "content": {
              "application/json": {
                "examples": {
                  "cancelled": {
                    "summary": "Sesi dibatalkan",
                    "value": {
                      "valid": true,
                      "session": {
                        "id": 1,
                        "code": "SES-K2M7-9QXW",
                        "status": "cancelled",
                        "mode": "self_service",
                        "started_at": "2026-09-13T04:10:00+00:00"
                      },
                      "cancelled": true,
                      "reason": null
                    }
                  },
                  "alreadyCancelled": {
                    "summary": "Sesi sudah dibatalkan sebelumnya (no-op)",
                    "value": {
                      "valid": true,
                      "session": {
                        "id": 1,
                        "code": "SES-K2M7-9QXW",
                        "status": "cancelled",
                        "mode": "self_service",
                        "started_at": "2026-09-13T04:10:00+00:00"
                      },
                      "cancelled": false,
                      "reason": "already_cancelled"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada atau bukan token perangkat"
          },
          "403": {
            "description": "Token tidak memiliki ability device:session atau perangkat telah dicabut"
          },
          "404": {
            "description": "Sesi tidak ditemukan atau bukan milik perangkat pemanggil"
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Profil Pengguna Terautentikasi",
        "description": "Mengambil detail profil pengguna yang sedang login berdasarkan Bearer Token.",
        "operationId": "getUserProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Detail User Berhasil Diambil",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "user": {
                    "id": 1,
                    "name": "Admin Photobooth Studio",
                    "email": "admin@photobooth.com"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}